extract the kernel lifecycle triggers from desktop
