+++ packages/core/src/epics/index.js
@@ -2,3 +2,8 @@
 export { executeCellEpic, updateDisplayEpic } from "./execute";
 export { commListenEpic } from "./comm";
 export { launchWebSocketKernelEpic } from "./websocket-kernel";
+
+export {
+  acquireKernelInfoEpic,
+  watchExecutionStateEpic
+} from "./kernel-lifecycle";
